meat is back on the menue boys!

+++ Color.cpp
@@ -1,6 +1,6 @@
 #pragma once
 struct Color {
-    int red;
-    int green;
-    int blue;
+    float red;
+    float green;
+    float blue;
 };
